@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdk 34
 
     defaultConfig {
         applicationId "fr.smartapps.android_sma_support_v1"
-        minSdkVersion 17
-        targetSdkVersion 27
+        minSdkVersion 21
+        targetSdkVersion 34
         versionCode 1
         versionName "1.0"
     }
@@ -19,9 +18,17 @@ android {
     }
 }
 
+repositories {
+    jcenter()
+    mavenCentral()
+    maven { url "https://jitpack.io" }
+    google()
+}
+
 dependencies {
     api project(':lib_commonlib')
     implementation 'com.github.vgachet:sma_assetmanager:1.0.36'
-    api 'com.android.support:percent:27.1.1'
-    api 'com.android.support:cardview-v7:27.1.1'
+    implementation 'com.github.smartapps-fr:TileView:1.0.0'
+    api 'androidx.percentlayout:percentlayout:1.0.0'
+    api 'androidx.cardview:cardview:1.0.0'
 }
