@@ -1,7 +1,7 @@
 package fr.smartapps.demo;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import fr.smartapps.lib.SMAAssetManager;
 import fr.smartapps.commonlib.imageview.SMAZoomableImageView;
