@@ -1,7 +1,7 @@
 package fr.smartapps.demo.video;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.DisplayMetrics;
 
 import fr.smartapps.demo.R;
