@@ -1,9 +1,7 @@
 package fr.smartapps.demo.video;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.View;
+import androidx.appcompat.app.AppCompatActivity;
 
 import fr.smartapps.demo.R;
 import fr.smartapps.lib.SMAAssetManager;
