@@ -1,6 +1,6 @@
 package fr.smartapps.demo.video;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import fr.smartapps.commonlib.video.SMAVideoView;
