@@ -2,7 +2,7 @@ package fr.smartapps.commonlib.pager;
 
 import android.app.Fragment;
 import android.app.FragmentManager;
-import android.support.v13.app.FragmentPagerAdapter;
+import androidx.legacy.app.FragmentPagerAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
